Code fix

+++ modules/numeral-system-calc/include/numeral_system_calc.h
@@ -8,13 +8,13 @@
 class Calc {
  private:
     static char _Priority(char a);
-    static long _Calculate(long num1, long num2, char operation);
-    static long _CalculateRpn(char *s);
+    static int _Calculate(int num1, int num2, char operation);
+    static int _CalculateRpn(char *s);
     static std::string _ConvertToRpn(const std::string &s);
 
  public:
-    static std::string toString(long number, size_t base);
-    static long calculate(const std::string &s);
+    static std::string toString(int number, size_t base);
+    static int calculate(const std::string &s);
 };
 
 #endif  // MODULES_NUMERAL_SYSTEM_CALC_INCLUDE_NUMERAL_SYSTEM_CALC_H_
