+++ modules/numeral-system-calc/test/test_numeral_system_calc.cpp
@@ -7,10 +7,10 @@
 TEST(Numeral_System_Calculate, Parse_Number) {
     // Arrange
     std::string s("030");  // 24
-    long answer = 24;
+    int answer = 24;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -19,10 +19,10 @@ TEST(Numeral_System_Calculate, Parse_Number) {
 TEST(Numeral_System_Calculate, Calculate_Binary) {
     // Arrange
     std::string s("1110011101001 + 101000001111");  // 7401 + 2575
-    long answer = 9976;
+    int answer = 9976;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -31,10 +31,10 @@ TEST(Numeral_System_Calculate, Calculate_Binary) {
 TEST(Numeral_System_Calculate, Calculate_Octal) {
     // Arrange
     std::string s("016351 + 05017");  // 7401 + 2575
-    long answer = 9976;
+    int answer = 9976;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -43,10 +43,10 @@ TEST(Numeral_System_Calculate, Calculate_Octal) {
 TEST(Numeral_System_Calculate, Calculate_Hexadecimal) {
     // Arrange
     std::string s("0x1CE9 + 0xA0F");  // 7401 + 2575
-    long answer = 9976;
+    int answer = 9976;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -55,10 +55,10 @@ TEST(Numeral_System_Calculate, Calculate_Hexadecimal) {
 TEST(Numeral_System_Calculate, Calculate_Different_Radices) {
     // Arrange
     std::string s("101 + 017 + 0xF");  // 5 + 15 + 15
-    long answer = 35;
+    int answer = 35;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -68,10 +68,10 @@ TEST(Numeral_System_Calculate, Calculate_Different_Radices) {
 TEST(Numeral_System_Calculate, Parse_Zero) {
     // Arrange
     std::string s("0 + 1 + 01 + 0 + 0xF + 0");  // 0 + 1 + 1 + 0 + 15 + 0
-    long answer = 17;
+    int answer = 17;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -80,10 +80,10 @@ TEST(Numeral_System_Calculate, Parse_Zero) {
 TEST(Numeral_System_Calculate, Different_Operators) {
     // Arrange
     std::string s("101 + 017 - 0xF * 10 / 05");  // 5 + 15 - 15*2 / 5
-    long answer = 14;
+    int answer = 14;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -92,10 +92,10 @@ TEST(Numeral_System_Calculate, Different_Operators) {
 TEST(Numeral_System_Calculate, Parentheses) {
     // Arrange
     std::string s("040 - (101 + 017 - 0xF) * 10 / 05");  // 32 - (5 + 15 - 15) * 2 / 5
-    long answer = 30;
+    int answer = 30;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
@@ -104,10 +104,10 @@ TEST(Numeral_System_Calculate, Parentheses) {
 TEST(Numeral_System_Calculate, Parentheses_Nested) {
     // Arrange
     std::string s("040 - (101 + (017 - 0xF) * 10) / 05");  // 32 - (5 + (15 - 15) * 2) / 5
-    long answer = 31;
+    int answer = 31;
 
     // Act
-    long res = Calc::calculate(s);
+    int res = Calc::calculate(s);
 
     // Assert
     EXPECT_EQ(answer, res);
